revamp: recipe's bulkactions in Bulkactions component
