+++ admin/src/shared/components/BulkAction/index.js
@@ -22,22 +22,85 @@ import {
    CONCATENATE_STRING_COLUMN,
    UPDATE_PRODUCT_OPTIONS,
 } from './mutation'
+import { RecipeBulkAction } from './entities/recipe'
 
 const BulkActions = ({
    children,
    table,
    selectedRows,
    removeSelectedRow,
-   bulkActions,
-   setBulkActions,
-   clearAllActions,
+   // bulkActions,
+   // setBulkActions,
+   setSelectedRows,
+   // clearAllActions,
    close,
-
    additionalBulkAction = {},
    additionalFunction,
 }) => {
+   // initial state for recipe
+   // need in bulk action component
+   const [initialBulkActionRecipe, setInitialBulkActionRecipe] = React.useState(
+      {
+         isPublished: false,
+         type: false,
+         cuisineName: {
+            defaultOption: null,
+            value: '',
+         },
+         author: '',
+         cookingTime: '',
+         utensils: '',
+         notIncluded: '',
+         description: '',
+         utensilsConcat: {
+            forAppend: '',
+            forPrepend: '',
+         },
+         notIncludedConcat: {
+            forAppend: '',
+            forPrepend: '',
+         },
+         descriptionConcat: {
+            forAppend: '',
+            forPrepend: '',
+         },
+      }
+   )
+   // need in bulk action component
+   const [bulkActions, setBulkActions] = React.useState({})
    const [showPopup, setShowPopup] = React.useState(false)
    const [popupHeading, setPopupHeading] = React.useState('')
+
+   // clear all actions
+   const clearAllActions = () => {
+      setInitialBulkActionRecipe(prevState => ({
+         ...prevState,
+         isPublished: !prevState.isPublished,
+         type: !prevState.type,
+         author: '',
+         cookingTime: '',
+         utensils: '',
+         cuisineName: {
+            defaultOption: null,
+            value: '',
+         },
+         notIncluded: '',
+         description: '',
+         utensilsConcat: {
+            forAppend: '',
+            forPrepend: '',
+         },
+         notIncludedConcat: {
+            forAppend: '',
+            forPrepend: '',
+         },
+         descriptionConcat: {
+            forAppend: '',
+            forPrepend: '',
+         },
+      }))
+      setBulkActions({})
+   }
    //mutation
    const [simpleRecipeUpdate] = useMutation(SIMPLE_RECIPE_UPDATE, {
       onCompleted: () => {
@@ -106,13 +169,10 @@ const BulkActions = ({
       switch (table) {
          case 'Recipe':
             return simpleRecipeUpdate
-            break
          case 'Product':
             return updateProducts
-            break
          case 'Ingredient':
             return updateIngredients
-            break
          case 'Product Options':
             return updateProductOptions
       }
@@ -276,6 +336,11 @@ const BulkActions = ({
                                  type="ghost"
                                  onClick={() => {
                                     removeSelectedRow(item.id)
+                                    setSelectedRows(prevState =>
+                                       prevState.filter(
+                                          row => row.id !== item.id
+                                       )
+                                    )
                                  }}
                               >
                                  <RemoveIcon color="#FF5A52" />
@@ -304,7 +369,15 @@ const BulkActions = ({
                   </Flex>
                   <Spacer size="8px" />
                   <Flex height="44vh" overflowY="auto">
-                     {children}
+                     {/* {children} */}
+                     {table === 'Recipe' && (
+                        <RecipeBulkAction
+                           initialBulkAction={initialBulkActionRecipe}
+                           setInitialBulkAction={setInitialBulkActionRecipe}
+                           bulkActions={bulkActions}
+                           setBulkActions={setBulkActions}
+                        />
+                     )}
                   </Flex>
                   <Spacer size="16px" />
                   <Flex container alignItems="center" justifyContent="flex-end">
