+++ admin/src/shared/components/BulkAction/subscription.js
@@ -1,0 +1,10 @@
+import gql from 'graphql-tag'
+
+export const CUISINES_NAMES = gql`
+   subscription Cuisines {
+      cuisineNames {
+         id
+         title: name
+      }
+   }
+`
